chore: move `MentionFormats` to `formats`
